pyproject.toml
[project]
name = "transcribe-translate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "moviepy>=2.1.2",
    "openai>=1.62.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
]
